+++ src/core/activate.ts
@@ -1,11 +1,11 @@
 import * as vscode from 'vscode';
-import { ExtViews } from '../types';
+import { ExtDeps } from '../types';
 import { registerCommands } from '../commands';
 import { extViews } from '../constants';
 import { registerWebviews } from '../webviews';
 import { setupSidebar } from '../sidebar';
 
-export const setupExt = (extViews: ExtViews, context: vscode.ExtensionContext) => {
+export const setupExt = (extViews: ExtDeps, context: vscode.ExtensionContext) => {
   registerCommands(context);
   registerWebviews(context);
   setupSidebar(context, vscode.workspace.workspaceFolders);