+++ src/sidebar/index.ts
@@ -1,4 +1,3 @@
-export * from './refreshViews';
 export * from './setupSidebar';
 export * from './setViewContext';
 export * from './shouldShowView';