+++ src/sidebar/setViewContext.ts
@@ -1,12 +1,12 @@
 import * as vscode from 'vscode';
 import { EXT } from '../constants';
-import { ExtViewList, GetPackageJsonResult } from '../types';
+import { ExtDepTypes, GetPackageJsonResult } from '../types';
 import { shouldShowView } from './';
 
 export const setViewContext = (view: string, packageJson: GetPackageJsonResult) => {
   vscode.commands.executeCommand(
     'setContext',
     `${EXT}-${view}`,
-    shouldShowView(view as ExtViewList, packageJson)
+    shouldShowView(view as ExtDepTypes, packageJson)
   );
 };