+++ src/sidebar/shouldShowView.ts
@@ -1,6 +1,6 @@
-import { GetPackageJsonResult, ExtViewList } from '../types';
+import { GetPackageJsonResult, ExtDepTypes } from '../types';
 
-export const shouldShowView = (view: ExtViewList, packageJson: GetPackageJsonResult): boolean => {
+export const shouldShowView = (view: ExtDepTypes, packageJson: GetPackageJsonResult): boolean => {
   if (packageJson !== null && !(packageJson instanceof Error)) {
     if (packageJson[view] !== undefined) {
       return true;