+++ src/types/ext.ts
@@ -1,13 +1,14 @@
 import { JSONSchemaForNPMPackageJsonFiles } from '@schemastore/package';
 import { FS_FOLDER_IMAGES_DARK, FS_FOLDER_IMAGES_LIGHT } from '../constants';
 import { PackageList, PackageListItem, PackageListDep } from '../treeviews';
+import { ExecException } from 'child_process';
 
-export type ExtViewList =
+export type ExtDepTypes =
   | 'dependencies'
   | 'devDependencies'
   | 'peerDependencies'
   | 'optionalDependencies';
-export type ExtViews = { [view in ExtViewList]: string };
+export type ExtDeps = { [view in ExtDepTypes]: string };
 
 export interface CmdCallbackData {
   packageName: string;
@@ -19,3 +20,12 @@ export type GetPackageJsonResult = JSONSchemaForNPMPackageJsonFiles | null | Err
 export type TreeProviders = { [key: string]: PackageList };
 export type PackageListChild = PackageListItem | PackageListDep;
 export type PackageListChildren = PackageListChild[];
+
+export type AddPackageData = {
+  dependencyType: ExtDepTypes;
+  error: ExecException | null;
+  stdout: string;
+  stderr: string;
+};
+
+export type AddPackage = Promise<AddPackageData>;